Partial data. Added code to get Train and Test Results
I added code to get the training and testing accuracies. The code take awhile to run so only part of the results are there. All the results will be avalible in a spreadsheet later. Our train/test ratio test would be more meaning full if we could test with datasets bigger than 50% of the dataset. See: https://github.com/thomaswsu/Stochastic-Gradient-Boosting-and-Adaboost/issues/2